+++ src/test/java/org/fwoxford/web/rest/FrozenTubeTypeResourceIntTest.java
@@ -49,8 +49,8 @@ public class FrozenTubeTypeResourceIntTest {
     private static final Integer DEFAULT_SAMPLE_USED_TIMES_MOST = 20;
     private static final Integer UPDATED_SAMPLE_USED_TIMES_MOST = 19;
 
-    private static final Integer DEFAULT_FROZEN_TUBE_VOLUMN = 20;
-    private static final Integer UPDATED_FROZEN_TUBE_VOLUMN = 19;
+    private static final Double DEFAULT_FROZEN_TUBE_VOLUMN = 20.0;
+    private static final Double UPDATED_FROZEN_TUBE_VOLUMN = 19.0;
 
     private static final String DEFAULT_FROZEN_TUBE_VOLUMN_UNIT = "AAAAAAAAAA";
     private static final String UPDATED_FROZEN_TUBE_VOLUMN_UNIT = "BBBBBBBBBB";
